macOS: Move MenuCommandRespondingViewController to a separate file
